# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CommentListResponse:
    properties:
      content:
        type: string
      created_at:
        $ref: '#/definitions/utils.JSONTime'
      deleted_at:
        $ref: '#/definitions/utils.JSONTime'
      id:
        type: integer
      parent_id:
        type: integer
      status:
        type: integer
      target:
        properties:
          key:
            type: string
          title:
            type: string
          type:
            type: string
        type: object
      user:
        properties:
          avatar:
            type: string
          badge:
            description: 铭牌
            type: string
          email:
            description: 后台显示邮箱
            type: string
          id:
            type: integer
          nickname:
            type: string
        type: object
    type: object
  dto.CommentResponse:
    properties:
      browser:
        description: 浏览器内核
        type: string
      content:
        type: string
      created_at:
        $ref: '#/definitions/utils.JSONTime'
      id:
        type: integer
      is_deleted:
        type: boolean
      location:
        description: 地理位置
        type: string
      os:
        description: 操作系统
        type: string
      parent_id:
        type: integer
      replies:
        items:
          $ref: '#/definitions/dto.CommentResponse'
        type: array
      reply_user:
        properties:
          avatar:
            type: string
          badge:
            description: 铭牌
            type: string
          email_hash:
            description: 邮箱哈希（用于 Gravatar）
            type: string
          id:
            type: integer
          nickname:
            type: string
          role:
            description: 角色
            type: string
          website:
            description: 网站
            type: string
        type: object
      user:
        properties:
          avatar:
            type: string
          badge:
            description: 铭牌
            type: string
          email_hash:
            description: 邮箱哈希（用于 Gravatar）
            type: string
          id:
            type: integer
          nickname:
            type: string
          role:
            description: 角色
            type: string
          website:
            description: 网站
            type: string
        type: object
    type: object
...
  dto.CreateCommentRequest:
    properties:
      content:
        maxLength: 500
        minLength: 1
        type: string
      email:
        type: string
      nickname:
        description: 游客信息（可选，未登录时必填）
        maxLength: 32
        minLength: 2
        type: string
      parent_id:
        type: integer
      target_key:
        type: string
      target_type:
        enum:
        - article
        - page
        type: string
      website:
        maxLength: 255
        type: string
    required:
    - content
    - target_key
    - target_type
    type: object
...
  dto.ImportCommentError:
    properties:
      content:
        type: string
      error:
        type: string
      index:
        type: integer
    type: object
  dto.ImportCommentsResult:
    properties:
      errors:
        items:
          $ref: '#/definitions/dto.ImportCommentError'
        type: array
      failed:
        type: integer
      success:
        type: integer
      total:
        type: integer
      user_created:
        type: integer
    type: object
...
  dto.UpdateCommentRequest:
    properties:
      content:
        maxLength: 500
        minLength: 1
        type: string
    required:
    - content
    type: object
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  utils.JSONTime:
    properties:
      time.Time:
        type: string
    type: object
...
info:
  contact:
    email: talen2004@163.com
    name: M.Talen
  description: 一个基于 Go 语言的现代化博客后端服务
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Flec-Server
  version: v1
paths:
...
  /admin/comments:
    get:
      consumes:
      - application/json
      description: 获取所有评论，支持按状态筛选
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量（不传则返回全部）
        in: query
        name: page_size
        type: integer
      - description: 状态筛选 0:隐藏 1:显示
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 评论列表
      tags:
      - 评论管理
    post:
      consumes:
      - application/json
      description: 管理员创建评论，用于回复用户，但无法设置评论IP等数据
      parameters:
      - description: 评论信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CommentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建评论
      tags:
      - 评论管理
  /admin/comments/{id}:
    delete:
      consumes:
      - application/json
      description: 软删除评论，子评论会保留，可通过恢复接口还原
      parameters:
      - description: 评论 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除评论
      tags:
      - 评论管理
    get:
      consumes:
      - application/json
      description: 查看评论详细信息
      parameters:
      - description: 评论 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CommentListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 评论详情
      tags:
      - 评论管理
  /admin/comments/{id}/restore:
    put:
      consumes:
      - application/json
      description: 恢复已删除的评论
      parameters:
      - description: 评论 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 恢复评论
      tags:
      - 评论管理
  /admin/comments/{id}/toggle-status:
    put:
      consumes:
      - application/json
      description: 切换评论的显示状态，隐藏后博客端不可见或被替换
      parameters:
      - description: 评论 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 切换评论状态
      tags:
      - 评论管理
  /admin/comments/import:
    post:
      consumes:
      - multipart/form-data
      description: 从Artalk评论系统导入评论数据，暂不支持更多评论系统
      parameters:
      - description: 导入文件（JSON或Artrans格式）
        in: formData
        name: file
        required: true
        type: file
      - description: 来源类型，目前支持：artalk
        enum:
        - artalk
        in: formData
        name: source_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ImportCommentsResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 导入评论数据
      tags:
      - 评论管理
...
  /comments:
    get:
      consumes:
      - application/json
      description: 获取目标评论，扁平化显示所有评论和回复。需指定 target_type 和 target_key
      parameters:
      - description: 目标类型(article/page)
        in: query
        name: target_type
        required: true
        type: string
      - description: 目标标识(文章slug或页面key)
        in: query
        name: target_key
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量（不传则返回全部）
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 评论列表
      tags:
      - 评论
    post:
      consumes:
      - application/json
      description: 发表评论或回复。已登录用户直接评论，未登录用户需提供昵称、邮箱等游客信息
      parameters:
      - description: 评论信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CommentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 发表评论
      tags:
      - 评论
  /comments/{id}:
    delete:
      consumes:
      - application/json
      description: 只能删除自己的评论，子评论会保留
      parameters:
      - description: 评论 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除评论
      tags:
      - 评论
    put:
      consumes:
      - application/json
      description: 只能修改自己的评论内容
      parameters:
      - description: 评论 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 评论信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CommentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 修改评论
      tags:
      - 评论
...
securityDefinitions:
  BearerAuth:
    description: 在请求头中添加 Bearer Token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"